requirements.txt
altair==4.2.2
appnope==0.1.3
arch==5.5.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
prompt-toolkit==3.0.38
property-cached==1.6.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==1.4.44
stack-data==0.6.2
statsmodels==0.14.0
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.3.1
traitlets==5.9.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
zipp==3.15.0
